+++ src/main/java/edu/aau/groupc/canteenbackend/user/dto/LoginDto.java
@@ -1,11 +1,11 @@
-package edu.aau.groupc.canteenbackend.dto;
+package edu.aau.groupc.canteenbackend.user.dto;
 
 import lombok.Data;
 
 import java.io.Serializable;
 
 @Data
-public class RegisterUserDto implements Serializable {
+public class LoginDto implements Serializable {
 
     private final String username;
     private final String password;
