+++ src/main/java/edu/aau/groupc/canteenbackend/user/dto/RegisterUserDto.java
@@ -1,11 +1,11 @@
-package edu.aau.groupc.canteenbackend.dto;
+package edu.aau.groupc.canteenbackend.user.dto;
 
 import lombok.Data;
 
 import java.io.Serializable;
 
 @Data
-public class LoginDto implements Serializable {
+public class RegisterUserDto implements Serializable {
 
     private final String username;
     private final String password;
